+++ spec/validator_spec.rb
@@ -357,6 +357,18 @@ RSpec.describe Evnt::Validator do
     expect(validation.passed?).to eq false
   end
 
+  it 'it should not accept a string that not match the regex option' do
+    value = 'hello'
+    validation = Evnt::Validator.new(value, type: :string, regex: /^(?=.*[a-zA-Z])(?=.*[0-9]).{6,20}$/)
+    expect(validation.passed?).to eq false
+  end
+
+  it 'it should accept a string that match the regex option' do
+    value = 'Hello1'
+    validation = Evnt::Validator.new(value, type: :string, regex: /^(?=.*[a-zA-Z])(?=.*[0-9]).{6,20}$/)
+    expect(validation.passed?).to eq true
+  end
+
   # Number:
   #################################################################